pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "eventgpt-scouting-app"
version = "0.1.0"
description = "Streamlit scouting engine UI on top of the eventGPT Modal service"
requires-python = ">=3.11"
dependencies = [
    "streamlit>=1.30",
    "httpx>=0.27",
    "openai>=1.40",
    "plotly>=5.22",
    "pandas>=2.0",
    "numpy>=1.24",
    "python-dotenv>=1.0",
    "modal>=1.0",
]

[project.optional-dependencies]
# Installed inside the Modal image only — torch + the scientific stack the
# inference-side library needs. The local Streamlit process does not need
# any of these.
modal-image = [
    "torch==2.4.1",
    "polars>=1.0",
    "pyarrow>=15",
    "pyyaml>=6",
    "pydantic>=2",
    "scikit-learn>=1.4",
    "hdbscan>=0.8.33",
    "umap-learn>=0.5",
    "fastapi>=0.110",
]

[tool.setuptools.packages.find]
include = ["app*", "eventgpt*"]
exclude = [".venv*", "configs*"]
